Split Kubacki Registry Number into sections
Changed the new calibration dialog so that the Kubacki registry number
is now split into separate sections representing different parts of the
complete number (a combination of text fields and labels).
The getFullKubackiRegistryNumber method from the dialog controller now
combines the values of those text fields and dots into a full title for
the new calibration tab.