Refactor variable and file names
Refactored most of the variable names used in the project so far to
their shorter counterparts. Also renamed a couple of class and fxml
files in the similar manner.
Changed the way the buttons' style from the main window toolbar is
modified: the styleClass property is now being used instead of the id
property.
Added a stub for the container class that is going to hold all the data
inputted into the new calibration dialog forms.