Add change listener to tab selectedProperty()
Ditched the whole idea of retrieving the sub-tabs of the new calibration
tab to attach a change listener to the selected property of the current
one for now.
Instead, the whole process is done in the CalTabController via directly
accessing the specific sub-tab controller reference.
Also, some general cleanup (removed test code, printouts etc.).