Refactor functionality for adding measurement ranges
Reimplemented container class for measurement data storage.
ValuationData is now used as a container for CalData that stores all the
information on ranges across all functions for that specific calibration
session.
TableViews for data presentation for all functions and the functionality
for adding extra frequencies for VAC and IAC work properly now.
Added some ContextMenu stubs for the basic 5 measurement functions
TableViews (items/buttons are not yet functional and incomplete).