Add initial implementation of vdc cost calculation
Just a stub to test if it works.
An observable property for VDC section cost added to the ValuationData
paired with a change listener attached to the observable list of VDC
ranges that recalculates the section's cost.
The cost is being displayed in the top panel of the section, above the
TableView. It's updated dynamically everytime the underlying property
changes.