Revamp the extra frequency adding functionality (WIP)
Started messing with the way extra VAC and IAC frequencies are being
added.
Extra frequencies are supposed to be stored as LinkedHashMaps -
frequency (String) as a key, and a list of measurement ranges as a
value.
Current work in progress is the ValuationData.removeExtraAcFreq
refactoring.
Also trying to implement a dialog for adding extra frequencies in
ValuationController.